requirements.txt
app==0.0.1
fastapi==0.78.0
numpy==1.22.4
Pillow==9.3.0
requests==2.27.1
scipy==1.9.3
starlette==0.19.1
uvicorn>=0.15.0,<0.18.0
pydantic>=1.8.0,<2.0.0


PyMatting==1.1.8
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
torchvision==0.12.0+cpu